renamed get_resonance_signal file
