+++ src/get_resonance_signal_CD.py
